fix issue with live reloading & speed up bundling process
The babel loader's exclude regex was not doing its job, so I think Webpack was scanning through node_modules to find .jsx files. This caused an error in the browser side live reloading code when serving the site with 'webpack-dev-server --hot --inline'. Swapping the exclude for an include fixes the js error & speeds up the bundling process.